ParetoRecipes

ParetoRecipes.pareto_front

ParetoRecipes.pareto_front — Function

Compute the Pareto front of a set of points.

Uses an inefficient brute-force exact method.

The Pareto front is also known as the Pareto frontier or Pareto set.